Move all classes from player.py to scripts/game_objects directory
This made me change the initialization methods in these classes to support modular architecture (e.g. player = Player(400, 300) or tile.display(screen))
